src/back/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
torch==2.3.0
numpy==1.26.4
pillow==10.3.0
